--- docker/dockerbuild
@@ -1,2 +1,0 @@
-cd $1
-docker build -t plebnet/$1 -f Dockerfile .
