--- docker/dockerrun
@@ -1,1 +1,0 @@
-docker run -it -p 127.0.0.1:8085:8085 plebnet/$1
