--- docker/market/README.md
@@ -1,1 +1,0 @@
-Setup tribler and the decentralized market.
